+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Feat: Add `graphql-java` instrumentation (#1777)
-* Fix: Do not crash when event processors throw a lower level Throwable class (#)
+* Fix: Do not crash when event processors throw a lower level Throwable class (#1800)
 
 ## 5.3.0