+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Feat: Add `graphql-java` instrumentation (#1777)
 * Fix: Do not crash when event processors throw a lower level Throwable class (#1800)
+* Fix: ActivityFramesTracker does not throw if Activity has no observers (#1799)
 
 ## 5.3.0